#4094400: 5x5 changes
